Subpixel rendering: iOS video playback controls look blurry. https://bugs.webkit.org/show_bug.cgi?id=135245 <rdar://problem/16878037>
Reviewed by Simon Fraser.

This patch introduces a compositing parent of the overlay control panel so that
the transformed overlay panel becomes sharp. This is a workaround for webkit.org/b/135246.

Can't find a way to test it yet.

Source/WebCore:
* Modules/mediacontrols/mediaControlsApple.css:
(video::-webkit-media-controls-panel-composited-parent):
* Modules/mediacontrols/mediaControlsApple.js:
(Controller.prototype.createControls):
(Controller.prototype.addControls):
* Modules/mediacontrols/mediaControlsiOS.css:
(video::-webkit-media-controls-panel-composited-parent):
* Modules/mediacontrols/mediaControlsiOS.js: This is a workaround for webkit.org/b/135248
It pushes the overlay panel down to close the gap with the video element. Since the
panel's size in css pixels is scale dependent, the gap needs to be scale dependent too.
(ControllerIOS.prototype.set pageScaleFactor):

LayoutTests:
* platform/mac/fast/hidpi/video-controls-in-hidpi-expected.txt:
* platform/mac/fast/layers/video-layer-expected.txt:
* platform/mac/media/audio-controls-rendering-expected.txt:
* platform/mac/media/controls-after-reload-expected.txt:
* platform/mac/media/controls-strict-expected.txt:
* platform/mac/media/controls-without-preload-expected.txt:
* platform/mac/media/media-controls-clone-expected.txt:
* platform/mac/media/video-no-audio-expected.txt:


Identifier: 153181@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@171593 268f45cc-cd09-0410-ab3c-d52691b4dbfc