[GTK] run-launcher --gtk still fails https://bugs.webkit.org/show_bug.cgi?id=135642
Patch by Philippe Normand <pnormand@igalia.com> on 2014-08-06
Reviewed by Carlos Garcia Campos.

The perl interpreter is confused by the combination of string
concatenation and a ternary in the same line. Using a separate
variable to determine the library file extension fixes this issue.

* Scripts/webkitdirs.pm:
(builtDylibPathForName): Use an intermediate variable, it's more
readable and unambiguous.

Identifier: 153509@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@172139 268f45cc-cd09-0410-ab3c-d52691b4dbfc