Fixed a bmalloc crash seen on the EWS bot https://bugs.webkit.org/show_bug.cgi?id=135955
Reviewed by Andreas Kling.

* bmalloc/Syscall.h: Some CG APIs vm_copy their input buffers. If the
input buffer is a malloc region, that region will get marked Copy-On-Write
by the kernel. Calls to madvise() for COW regions fail and return EINVAL
on older OS X's. In 10.10, they still fail, but they do not return
EINVAL.

So, we can only ASSERT that our syscalls succeed starting with 10.10.


Identifier: 153813@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@172608 268f45cc-cd09-0410-ab3c-d52691b4dbfc